Feature/premake linux (#689)
* Premake work in progress for linux.

* Added dump function.

* Remove examples on linux
Small warning fix.

* * Don't build render-test on linux
* Removed work around virtual destructor warning, and just used virtual dtor for simplicity
* Git ignore obj directories

* Fix premake working on windows.

* * Fix sprintf_s functions
* Make generates arg parsing more robust

* Added FloatIntUnion to avoid type punning/strong aliasing issues, and repeated union definitions.

* Work around problems building on linux with getClass claiming a strict aliasing issue.

* Fix for targetBlock appearing potentiall used unintialized to gcc.

* Linux slang link options -fPIC to make dll.

* Add -fPIC to build options on linux.

* Add -ldl for linux on slang.

* Fixes to try and get premake working with .so on linux.

* Make core compile with -fPIC

* Try to fix linux linking with --no-as-needed before -ldl

* Add rpath back.

* Remove render-gl from linux build.

* Re-add location for linux.

* Don't include <malloc.h> except on windows.

* Remove unused line to fix warning on osx.

* Remove ambiguity on OSX for operator <<.

* Fixing ambiguity with operator overloading and Int types for OSX.

* Fix ambiguity around UInt and operator

* Fix ambiguity of UInt conversion for OSX.

* Added UnambiguousInt and UnambiguousUInt to make it easier to work around OSX integer coercion for UInt/Int types.

+++ source/slang/ir-legalize-types.cpp
@@ -1201,7 +1201,7 @@ static LegalVal declareSimpleVar(
                 if (mangledNameStr.Length() != 0)
                 {
                     mangledNameStr.append("L");
-                    mangledNameStr.append(globalNameInfo->counter++);
+                    mangledNameStr.append(Int32(globalNameInfo->counter++));
                     globalVar->mangledName = context->session->getNameObj(mangledNameStr);
                 }
             }
